requirements.txt
datetime
numpy
openpyxl
pandas
pdfminer.six
python-dotenv
xlrd
